+++ miso-web/src/main/webapp/scripts/hot_sample.js
@@ -4,10 +4,13 @@
  */
 HotTarget.sample = (function() {
 
-  var getSampleClass = function(sample) {
+  var getSampleCategory = function(sample) {
+    if (!Constants.isDetailedSample) {
+      return 'Plain';
+    }
     return Constants.sampleClasses.find(function(sampleClass) {
       return sample.sampleClassId == sampleClass.id;
-    });
+    }).sampleCategory;
   };
 
   var getSampleClasses = function(samples) {
@@ -1178,11 +1181,13 @@ HotTarget.sample = (function() {
             }
           },
           HotUtils.printAction('sample'),
-          HotUtils.spreadsheetAction(Urls.rest.samples.spreadsheet, Constants.sampleSpreadsheets, function(samples, spreadsheet) {
+          HotUtils.spreadsheetAction(Urls.rest.samples.spreadsheet, Constants.sampleSpreadsheets.filter(function(sheet) {
+            return Constants.isDetailedSample || sheet.allowedClasses.indexOf('Plain') !== -1;
+          }), function(samples, spreadsheet) {
             var errors = [];
             var invalidSamples = [];
             samples.forEach(function(sample) {
-              if (!spreadsheet.sheet.allowedClasses.includes(getSampleClass(sample).sampleCategory)) {
+              if (!spreadsheet.sheet.allowedClasses.includes(getSampleCategory(sample))) {
                 invalidSamples.push(sample);
               }
             })
@@ -1191,7 +1196,7 @@ HotTarget.sample = (function() {
               errors.push("Allowed types: " + spreadsheet.sheet.allowedClasses.join(", "));
               errors.push("Invalid samples:")
               invalidSamples.forEach(function(sample) {
-                errors.push("* " + sample.alias + " (" + getSampleClass(sample).alias + ")");
+                errors.push("* " + sample.alias + " (" + getSampleCategory(sample) + ")");
               })
             }
             return errors;